Fix layout grid in site editor.
This PR fixs #154 where the layout grid did not work in the site editor. The rules were simply not output, because https://github.com/Automattic/block-experiments/commit/4ac3fb6166024cc447725b8c5a5883b4cadd2c89 targets a rule that exists only in the post editor.

This PR hopes to keep the intent of that commit in place — scoping the rules only to the block editor — but target a class that exists in both the site and post editor.